RF: move for newer images to use C.UTF-8 not en_US.UTF-8
This is somewhat inspired by kwik image blowing with

   abort further execution because Python 3 was configured to use ASCII as encoding for the environment. Consult https://click.palletsprojects.com/en/7.x/python3/ for mitigation steps.

This system supports the C.UTF-8 locale which is recommended.
You might be able to resolve your issue by exporting the
following environment variables:

    export LC_ALL=C.UTF-8
    export LANG=C.UTF-8

Click discovered that you exported a UTF-8 locale
but the locale system could not pick up from it because
it does not exist.  The exported locale is "en_US.UTF-8" but it
is not supported
scripts/singularity_cmd run  --help  3.82s user 2.43s system 118% cpu 5.289 total

so it feels that C. instead of en_US. could be a safer default