BF: adjusted sed rule for parsing docker tags
I have noted that we got a bunch of images with "v" prefix in the singularity
image tag, so something got off in our version parsing.  Also sed rule
might have been succeptable to have jumping group due to ()? and there
seemed to be not needed parantheses.  So with this change the first
nested group (\2) for the possible prefix is always there but could be
empty. And overall match (\1) will be used for the entire version