BF(DOC): use ///repronim/containers directly + provide complete script working in TMPDIR
original report: https://github.com/ReproNim/containers/issues/44

Use /// since that is where we have all the images and there is a bug to be fixed on datalad
end: https://github.com/datalad/datalad/issues/2897

With using TMPDIR and () we can make it cut/pasteable and/or script.  We could
possibly even automate testing of this snippet since it would be easy to
identify it in the README.md (might do later).  And set -eu  is a demonstration
of good practices we try to encourage in our training materials